requirements.txt
asgiref==3.8.1
astroid==3.3.8
colorama==0.4.6
dill==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
isort==6.0.1
mccabe==0.7.0
pillow==11.1.0
platformdirs==4.3.6
psycopg2==2.9.10
PyJWT==2.9.0
pylint==3.3.4
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
sqlparse==0.5.3
tomlkit==0.13.2
tzdata==2025.1

celery~=5.5.3
drf-yasg~=1.21.10